fix: moved subNavigationItem to partials
