+++ src/components/common/Extensions/SubNavigation/SubNavigationMobile.tsx
@@ -2,7 +2,7 @@ import React, { PropsWithChildren, useState, FC } from 'react';
 
 import { useDetectClickOutside } from '~hooks';
 
-import SubNavigationItemMobile from './SubNavigationItem/SubNavigationItemMobile';
+import SubNavigationItemMobile from './partials/SubNavigationItem/SubNavigationItemMobile';
 import { subNavigationItems } from './consts';
 import styles from './SubNavigationMobile.module.css';
