+++ src/components/common/Extensions/SubNavigation/partials/SubNavigationItem/SubNavigationItemMobile.tsx
@@ -10,7 +10,7 @@ import { SubNavigationItemProps } from './types';
 import styles from './SubNavigationItemMobile.module.css';
 import { accordionAnimation } from '~constants/accordionAnimation';
 
-const displayName = 'common.Extensions.SubNavigation.SubNavigationItem.SubNavigationItemMobile';
+const displayName = 'common.Extensions.SubNavigation.partials.SubNavigationItem.SubNavigationItemMobile';
 
 const SubNavigationItemMobile: FC<PropsWithChildren<SubNavigationItemProps>> = ({
   label,
